Move harvest collectors into harvest/harvesters.
Rename harvests modules to singular collector files and update tasks and tests to the new import paths.
